+++ AAEmu.Game/Utils/Scripts/SubCommands/Time/TimeSetSubCommand.cs
@@ -19,7 +19,7 @@ public class TimeSetSubCommand : SubCommandBase
         var oldTime = TimeManager.Instance.GetTime;
         var hour = (int)oldTime;
         if (parameters.TryGetValue("hour", out var hourValue))
-            hour = (int)hourValue;
+            hour = hourValue;
         var minute = (int)Math.Floor((oldTime - Math.Truncate(oldTime)) * 60f);
         if (parameters.TryGetValue("minute", out var minuteValue))
             minute = minuteValue;
